requirements.txt
# Core web framework
fastapi==0.115.12
uvicorn==0.34.2

# HTTP client libraries
aiohttp==3.12.0
requests==2.32.3

# Authentication and Azure libraries
msal==1.32.3

# Environment and configuration
python-dotenv==1.1.0

# Templating
Jinja2==3.1.6

# Background tasks and form handling
python-multipart

# Static files for web UI
fastapi-static-files==1.0.1

# Data processing and analysis
pandas==2.2.3
numpy>=1.26.0  # Compatible with Python 3.13

# Optional: Enhanced logging and monitoring
structlog==23.2.0

# Development dependencies (optional)
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
